更新
v0.0.2.0
2019-10-26

+++ Events/LunarEclipse.cs
@@ -12,18 +12,13 @@ namespace DisorderUnderstar.Events
 {
     public class LunarEclipse
     {
-        private static bool 持续事件 = false;
-        public static bool 每夜晚检测 = true;
+        public static int 杀怪分数 = 0;
+        public static bool 事件发生中 = false;
         public static bool 使用月蚀耀碑 = false;
-        #region 难度递增
-        public static bool 打败世纪之花 = false;
-        public static bool 打败月球领主 = false;
-        public static bool 打败月球领主后克鲁苏之眼 = false;
-        public static bool 打败月球领主后血肉之墙 = false;
-        public static bool 打败月球领主后世纪之花 = false;
-        #endregion
+        private int 波数 = 0;
+        private static bool 每夜晚检测 = true;
         #region 特定怪物生成
-        private static int 生成_世纪之花;
+        private int 生成_月耀之眼;
         #endregion
         public static bool 符合事件出现条件_月蚀()
         {
@@ -40,37 +35,35 @@ namespace DisorderUnderstar.Events
             else if (!每夜晚检测 && Main.dayTime) { 每夜晚检测 = true; }
             return !Main.dayTime && !Main.bloodMoon && Main.hardMode && !Main.pumpkinMoon && !Main.snowMoon && !DD2Event.Ongoing && 概率选中;
         }
-        public static bool 判定(Mod mod, NPC target, Item item, Player player, Projectile projectile)
+        public bool 判定(NPC target, Item item, Player player, Projectile projectile)
         {
-            if (Main.dayTime)
+            if (Main.dayTime || !完成事件())
             {
-                持续事件 = false;
+                事件发生中 = false;
                 使用月蚀耀碑 = false;
+                return false;
             }
-            if (使用月蚀耀碑 || 符合事件出现条件_月蚀_有概率() || 持续事件)
+            if (使用月蚀耀碑 || 符合事件出现条件_月蚀_有概率() || 事件发生中)
             {
-                持续事件 = true;
-                事件发生(mod, target, item, player, projectile);
+                事件发生中 = true;
+                事件发生(item);
+                事件发生(player);
+                事件发生(target);
+                事件发生(projectile);
             }
-            return 持续事件;
+            return 事件发生中;
         }
-        public static void 事件发生(Mod mod, NPC target, Item item, Player player, Projectile projectile)
+        private bool 完成事件()
         {
-            if (target.active)
-            {
-                if (target.friendly)
-                {
-                    target.velocity *= 0.5f;
-                    target.knockBackResist = 5f;
-                }
-            }
+            if (波数 > AllType.月蚀_杀怪分数.Length) { return true; }
+            else if (杀怪分数 >= AllType.月蚀_杀怪分数[波数]) { 波数 += 1; }
+            return false;
+        }
+        public void 事件发生(Item item)
+        {
             if (item.accessory)
             {
-                if (item.defense > 10)
-                {
-                    item.value = Item.sellPrice(0, 0, 50, 0);
-                    item.defense = 10;
-                }
+                if (item.defense > 10) { item.defense = 10; }
             }
             if (item.melee)
             {
@@ -78,7 +71,7 @@ namespace DisorderUnderstar.Events
                 if (item.damage > item.crit * 5) { item.damage = item.crit * 5; }
                 if (item.useTime < item.useAnimation) { item.useTime = item.useAnimation; }
                 if (item.useAnimation < 30) { item.useAnimation = 30; }
-                ItemOverride.事件_月蚀(item);
+                // ItemOverride.事件_月蚀(item);
             }
             if (item.magic)
             {
@@ -90,33 +83,40 @@ namespace DisorderUnderstar.Events
             }
             if (item.ranged)
             {
-                Projectile pOWNER = Main.projectile[item.shoot];
-                if (item.useAmmo == AmmoID.Arrow &&
-                    (pOWNER.Name.Contains("Arrow") || pOWNER.Name.Contains("箭") || pOWNER.Name.Contains("矢")) &&
-                    projectile.owner == item.owner)
+                Projectile projectile = Main.projectile[item.shoot];
+                if (projectile.Name.Contains("Arrow") || projectile.Name.Contains("箭") || projectile.Name.Contains("矢"))
                 { item.shoot = ProjectileID.WoodenArrowFriendly; }
-                else if (item.useAmmo == AmmoID.Bullet &&
-                    (pOWNER.Name.Contains("Bullet") || pOWNER.Name.Contains("子弹")) &&
-                    projectile.owner == item.owner)
+                else if (projectile.Name.Contains("Bullet") || projectile.Name.Contains("子弹"))
                 { item.shoot = ProjectileID.Bullet; }
-                else if (item.useAmmo == AmmoID.Coin &&
-                    (pOWNER.Name.Contains("Coin") || pOWNER.Name.Contains("币") || pOWNER.Name.Contains("钱")) &&
-                    projectile.owner == item.owner)
+                else if (projectile.Name.Contains("Coin") || projectile.Name.Contains("币") || projectile.Name.Contains("钱"))
                 { item.shoot = ProjectileID.CopperCoin; }
             }
+            return;
+        }
+        public void 事件发生(Player player)
+        {
             if (player.active)
             {
                 player.nightVision = true;
                 if (player.statDefense > 128) { player.statDefense = 128; }
             }
-            if (projectile.active && projectile.friendly &&
-                (projectile.type != mod.ProjectileType<ProStarArrow>() || projectile.type != mod.ProjectileType<ProWtfway>()))
+            if (NPC.downedPlantBoss && 波数 >= 5) { 生成怪物_世纪之花后(); }
+        }
+        public void 事件发生(NPC target)
+        {
+            if (target.active)
             {
-                projectile.penetrate = 1;
-                projectile.extraUpdates += 2;
-                if (projectile.damage > 0) { projectile.damage = 0; }
-                if (projectile.knockBack > 5f) { projectile.knockBack = 5f; }
+                if (target.friendly)
+                {
+                    target.velocity *= 0.8f;
+                    target.knockBackResist = 5f;
+                }
             }
+            return;
+        }
+        public void 事件发生(Projectile projectile)
+        {
+            Player player = Main.player[Main.myPlayer];
             if (projectile.active && projectile.hostile)
             {
                 float disP = Vector2.Distance(projectile.Center, player.Center);
@@ -125,7 +125,7 @@ namespace DisorderUnderstar.Events
                 {
                     projectile.Kill();
                     player.statLife -= 100;
-                    player.AddBuff(mod.BuffType<DebuffLunarErosion>(), 60);
+                    player.AddBuff(ModContent.BuffType<DebuffLunarErosion>(), 60);
                 }
                 else
                 {
@@ -142,9 +142,9 @@ namespace DisorderUnderstar.Events
                         projectile.Kill();
                         npc.life -= 100;
                         npc.lifeMax -= npc.lifeMax / 10;
-                        npc.AddBuff(mod.BuffType<DebuffLunarErosion>(), 120);
+                        npc.AddBuff(ModContent.BuffType<DebuffLunarErosion>(), 120);
                     }
-                    else if (!npc.friendly && npc.active && disMAX >= disN)
+                    else if (!npc.friendly && npc.active && Main.rand.Next(1, 100) <= 1 && disMAX >= disN)
                     {
                         projectile.Kill();
                         npc.life += npc.lifeMax / 2;
@@ -152,16 +152,42 @@ namespace DisorderUnderstar.Events
                     }
                 }
             }
-            if (打败世纪之花) { 生成怪物_世纪之花后(mod, player); }
-        }
-        public static void 生成怪物_世纪之花后(Mod mod, Player player)
-        {
-            if (Main.rand.Next(1, 1000) <= 1 && 生成_世纪之花 <= 2)
+            foreach (NPC target in Main.npc)
             {
-                NPC.NewNPC((int)player.Center.X + Main.screenWidth + Main.rand.Next(-30, 30), (int)player.Center.Y + Main.rand.Next(-30, 30),
-                    mod.NPCType<暂定名_以后要改>());
-                生成_世纪之花 += 1;
+                if (projectile.active && projectile.friendly)
+                {
+                    float dis = Vector2.Distance(projectile.Center, target.Center);
+                    float disMAX = 0.1f;
+                    projectile.penetrate = 1;
+                    projectile.extraUpdates += 2;
+                    if (projectile.knockBack > 3f) { projectile.knockBack = 3f; }
+                    if (disMAX >= dis && projectile.CanHit(target)) { projectile.damage = target.lifeMax / 8; }
+                }
             }
+            return;
+        }
+        private void 生成怪物_世纪之花后()
+        {
+            Player player = Main.player[Main.myPlayer];
+            #region 月耀之眼
+            if (Main.rand.Next(1, 1000) <= 1 && 生成_月耀之眼 <= 2)
+            {
+                if (Main.rand.NextBool())
+                {
+                    int yPosition = (int)player.Center.Y + Main.rand.Next(-20, 20);
+                    int xPosition = (int)player.Center.X + Main.screenWidth + Main.rand.Next(-20, 20);
+                    NPC.NewNPC(xPosition, yPosition, ModContent.NPCType<EyeofLunarShine>());
+                }
+                else
+                {
+                    int yPosition = (int)player.Center.Y + Main.rand.Next(-20, 20);
+                    int xPosition = (int)player.Center.X - Main.screenWidth + Main.rand.Next(-20, 20);
+                    NPC.NewNPC(xPosition, yPosition, ModContent.NPCType<EyeofLunarShine>());
+                }
+                生成_月耀之眼 += 1;
+                return;
+            }
+            #endregion
         }
     }
 }
